# swagger-static/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/users": {
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "cookie": []
          }
        ]
      }
    },
    "/users/profile": {
      "get": {
        "operationId": "UsersController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get profile information successfully"
          },
          "403": {
            "description": "Forbidden."
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "cookie": []
          },
          {
            "bearer": []
          }
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_signIn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInDto"
              },
              "examples": {
                "user_1": {
                  "value": {
                    "phoneNumber": "0898641337",
                    "password": "123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "AuthController_signUp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthDto"
              },
              "examples": {
                "user_1": {
                  "value": {
                    "phoneNumber": "0123456789",
                    "fullName": "Nguyen Van A",
                    "gender": "male",
                    "dob": "01/01/2024",
                    "password": "123",
                    "rePassword": "123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Register successfully!"
          },
          "409": {
            "description": "Existed."
          }
        },
        "tags": [
          "auth"
        ]
      }
    }
  },
  "info": {
    "title": "Bike Go",
    "description": "The BikeGo API description",
    "version": "1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "BikeGo",
      "description": ""
    }
  ],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {}
      },
      "SignInDto": {
        "type": "object",
        "properties": {
          "phoneNumber": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "phoneNumber",
          "password"
        ]
      },
      "AuthDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}